requirements.txt
pandas
streamlit
gspread
Pillow